requirements.txt
astroid==2.3.2
autopep8==1.4.4
blis==0.4.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.8
cymem==2.0.2
Django==2.2.6
django-rest-knox==4.1.0
django-webpack-loader==0.6.0
djangorestframework==3.10.3
en-core-web-sm==2.2.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
joblib==0.13.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.2
optional-django==0.1.0
plac==0.9.6
postgres==3.0.0
preshed==3.0.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
selenium==3.141.0
six==1.12.0
sklearn==0.0
spacy==2.2.2
sqlparse==0.3.0
srsly==0.1.0
stop-words==2018.7.23
stopwords==0.1.3
thinc==7.3.1
tqdm==4.36.1
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.5
wasabi==0.4.0
wrapt==1.11.2
zipp==0.6.0
